+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADB2BManagementPolicy/MSFT_AADB2BManagementPolicy.psm1
@@ -1,3 +1,5 @@
+Confirm-M365DSCModuleDependency -ModuleName 'MSFT_AADB2BManagementPolicy'
+
 function Get-TargetResource
 {
     [CmdletBinding()]
@@ -382,3 +384,4 @@ function Export-TargetResource
 }
 
 Export-ModuleMember -Function *-TargetResource
+