+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneCloudProvisioningPolicyWindows365/MSFT_IntuneCloudProvisioningPolicyWindows365.psm1
@@ -1,3 +1,5 @@
+Confirm-M365DSCModuleDependency -ModuleName 'MSFT_IntuneCloudProvisioningPolicyWindows365'
+
 function Get-TargetResource
 {
     [CmdletBinding()]
@@ -890,3 +892,4 @@ function Get-CompareParameters
 }
 
 Export-ModuleMember -Function @('*-TargetResource', 'Get-CompareParameters')
+