+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneRoleAssignmentWindows365/MSFT_IntuneRoleAssignmentWindows365.psm1
@@ -1,3 +1,5 @@
+Confirm-M365DSCModuleDependency -ModuleName 'MSFT_IntuneRoleAssignmentWindows365'
+
 function Get-TargetResource
 {
     [CmdletBinding()]
@@ -626,3 +628,4 @@ function Export-TargetResource
 }
 
 Export-ModuleMember -Function *-TargetResource
+