+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneTermsAndConditions/MSFT_IntuneTermsAndConditions.psm1
@@ -1,3 +1,5 @@
+Confirm-M365DSCModuleDependency -ModuleName 'MSFT_IntuneTermsAndConditions'
+
 function Get-TargetResource
 {
     [CmdletBinding()]
@@ -546,3 +548,4 @@ function Export-TargetResource
 }
 
 Export-ModuleMember -Function *-TargetResource
+